# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/ops/bookings/{bookingId}/tasks": {
      "get": {
        "operationId": "TaskController_list",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "tasks"
        ]
      },
      "post": {
        "operationId": "TaskController_create",
        "parameters": [
          {
            "name": "bookingId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/ops/tasks/{taskId}": {
      "patch": {
        "operationId": "TaskController_update",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/ops/tasks/{taskId}/status": {
      "post": {
        "operationId": "TaskController_status",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskStatusDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
    "/api/v1/ops/tasks/{taskId}/comment": {
      "post": {
        "operationId": "TaskController_comment",
        "parameters": [
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskCommentDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "tags": [
          "tasks"
        ]
      }
    },
...
  },
  "info": {
    "title": "Foresee Reservation API",
    "description": "Booking and operations API for Foresee Corporation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    },
    "schemas": {
...
      "CreateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "โทรยืนยันข้อมูลกับลูกค้า"
          },
          "description": {
            "type": "string",
            "example": "ยืนยันปริมาณและเวลาเข้าบริการ"
          },
          "assigneeUserId": {
            "type": "string",
            "format": "uuid"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "URGENT"
            ],
            "default": "NORMAL"
          },
          "dueAt": {
            "type": "string",
            "example": "2026-09-10T10:00:00.000Z"
          },
          "isRequired": {
            "type": "boolean",
            "default": false
          }
        },
        "required": [
          "title"
        ]
      },
      "UpdateTaskDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "โทรยืนยันข้อมูลกับลูกค้า"
          },
          "description": {
            "type": "string",
            "example": "ยืนยันปริมาณและเวลาเข้าบริการ"
          },
          "assigneeUserId": {
            "type": "string",
            "format": "uuid"
          },
          "priority": {
            "type": "string",
            "enum": [
              "LOW",
              "NORMAL",
              "HIGH",
              "URGENT"
            ],
            "default": "NORMAL"
          },
          "dueAt": {
            "type": "string",
            "example": "2026-09-10T10:00:00.000Z"
          },
          "isRequired": {
            "type": "boolean",
            "default": false
          },
          "version": {
            "type": "number",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "WAITING",
              "BLOCKED",
              "DONE",
              "CANCELLED"
            ]
          },
          "changeReason": {
            "type": "string",
            "example": "ลูกค้าแจ้งข้อมูลครบแล้ว"
          }
        },
        "required": [
          "title",
          "version"
        ]
      },
      "TaskStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "TODO",
              "IN_PROGRESS",
              "WAITING",
              "BLOCKED",
              "DONE",
              "CANCELLED"
            ]
          },
          "reason": {
            "type": "string",
            "example": "รอเอกสารจากลูกค้า"
          }
        },
        "required": [
          "status"
        ]
      },
      "TaskCommentDto": {
        "type": "object",
        "properties": {
          "note": {
            "type": "string",
            "example": "โทรแจ้งลูกค้าแล้ว รอส่งรูปหน้างาน"
          }
        },
        "required": [
          "note"
        ]
      }
    }
  }
}