# packages/api-client/src/generated/model/createTaskDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Foresee Reservation API
 * Booking and operations API for Foresee Corporation
 * OpenAPI spec version: 1.0
 */
import type { CreateTaskDtoPriority } from './createTaskDtoPriority';

export interface CreateTaskDto {
  title: string;
  description?: string;
  assigneeUserId?: string;
  priority?: CreateTaskDtoPriority;
  dueAt?: string;
  isRequired?: boolean;
}