# packages/api-client/src/generated/model/taskCommentDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Foresee Reservation API
 * Booking and operations API for Foresee Corporation
 * OpenAPI spec version: 1.0
 */

export interface TaskCommentDto {
  note: string;
}