# packages/api-client/src/generated/model/taskStatusDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Foresee Reservation API
 * Booking and operations API for Foresee Corporation
 * OpenAPI spec version: 1.0
 */
import type { TaskStatusDtoStatus } from './taskStatusDtoStatus';

export interface TaskStatusDto {
  status: TaskStatusDtoStatus;
  reason?: string;
}