# packages/api-client/src/generated/model/updateTaskDto.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Foresee Reservation API
 * Booking and operations API for Foresee Corporation
 * OpenAPI spec version: 1.0
 */
import type { UpdateTaskDtoPriority } from './updateTaskDtoPriority';
import type { UpdateTaskDtoStatus } from './updateTaskDtoStatus';

export interface UpdateTaskDto {
  title: string;
  description?: string;
  assigneeUserId?: string;
  priority?: UpdateTaskDtoPriority;
  dueAt?: string;
  isRequired?: boolean;
  version: number;
  status?: UpdateTaskDtoStatus;
  changeReason?: string;
}